Chatbot/requirements.txt
# Chatbot Backend Dependencies

fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1

# Vector Database - Qdrant (pre-built wheels, no C++ compilation)
qdrant-client==1.12.1

# LangChain & LangGraph for RAG
langchain==0.3.14
langchain-community==0.3.14
langgraph==0.2.62

# Groq API for LLM
groq==0.13.1
langchain-groq==0.2.2

# Embeddings & Re-ranking
sentence-transformers==3.3.1
# Cross-encoder re-ranker (ms-marco-MiniLM-L-6-v2) is downloaded automatically
# via sentence-transformers — no extra package needed

# torch is a sentence-transformers transitive dep; pin CPU build for servers
# without a GPU to keep the image small (override with torch+cu121 for GPU)
torch>=2.0.0

# Sentiment Analysis
transformers==4.47.1

# Web Scraping
beautifulsoup4==4.12.3
lxml==5.3.0
selenium==4.27.1
webdriver-manager==4.0.2

# Utilities
pydantic==2.10.4
httpx==0.28.1
numpy
